% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{basicDescr}
\alias{basicDescr}
\title{Basic descriptive statistics}
\usage{
basicDescr(x)
}
\arguments{
\item{x}{a numeric variable}
}
\value{
a data.frame
}
\description{
Return mean, standard deviation, skewness, and kurtosis of a variable.
}

