% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{classifyItemsMirt}
\alias{classifyItemsMirt}
\title{Classify Items for MIRT}
\usage{
classifyItemsMirt(fm, Df, min.loading = 0.4, max.loading = 0.3,
  max.complexity = 100, itemlength = 0, digits = 2,
  Df.name = deparse(substitute(Df)))
}
\arguments{
\item{fm}{a factor model fitted by mirt::mirt}

\item{Df}{a data frame containing the items}

\item{min.loading}{minimum loading of an item to be considered a marker of a factor}

\item{max.loading}{maximum loading of an item on a secondary factor (i.e., the factor on which the items has its second highest loading) to be considered a marker for the primary factor (i.e., the factor on which the items has its highest loading)}

\item{max.complexity}{maximum complexity of an item to be considered a marker item}

\item{itemlength}{trim item text to given number characters (0 = automatic trimming)}

\item{digits}{number of digits used in the output}

\item{Df.name}{name of data frame}
}
\description{
Automatically classify the items in a data frame.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}
