% Generated by roxygen2 (4.0.1): do not edit by hand
\name{descriptives}
\alias{descriptives}
\title{Descriptive Statistics of the Variables in a Data Frame}
\usage{
descriptives(x, width = 8, digits = 2)
}
\arguments{
\item{x}{a data frame}

\item{width}{width of numbers}

\item{digits}{preferred number of digits after the decimal point}
}
\description{
\code{descriptives} lists means, standard deviations, minima, maxima, number of missing values, and number of non-missing (valid) values of the numeric variables in a data frame. It aims at giving a quick overview of a data frame containing test/questionnaire items.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

