% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{factoranalysis}
\alias{factoranalysis}
\title{Exploratory Factor Analysis}
\usage{
factoranalysis(x, nfactors, rotate = "promax", fm = "ml")
}
\arguments{
\item{x}{a data frame}

\item{nfactors}{number of factors to be extracted}

\item{rotate}{rotation}

\item{fm}{factoring method}
}
\description{
Perform exploratory factor analysis of a set of variables. This is a wrapper around psych::fa.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

