\name{getDataFrames}
\alias{getDataFrames}
\title{Data frames in global workspace}
\usage{
  getDataFrames()
}
\value{
  A character vector with the names of the data frames
}
\description{
  Returns of a vector with the names of the data frames
  present in the global environment.
}
\author{
  Michael Hock \email{michael.hock@uni-bamberg.de}
}

