% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDataFramesIana}
\alias{getDataFramesIana}
\title{Data frames in global workspace}
\usage{
getDataFramesIana()
}
\value{
A character vector with the names of the data frames
}
\description{
Returns of a vector with the names of the data frames present in the
global environment. If no data frame is found, Iana's example data
are loaded.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

