% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{ggplotICC.RM}
\alias{ggplotICC.RM}
\title{ICCs for Rasch Model}
\usage{
ggplotICC.RM(object, empICC = NULL, empCI = NULL, xlim = c(-4, 4),
  xlab = "Dimension", ylab = "Probability")
}
\arguments{
\item{object}{of class RM}

\item{empICC}{see eRm::plotICC}

\item{empCI}{not used yet}

\item{xlim}{see eRm::plotICC}

\item{xlab}{see eRm::plotICC}

\item{ylab}{see eRm::plotICC}
}
\description{
Plot item characteristic curves for a Rasch model.
}
\details{
This a slightly modified version of plotICC.Rm from package eRm (version 0.15-1). The main  difference is the usage of ggplot2 for plotting.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

