% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{ggscree.plot}
\alias{ggscree.plot}
\title{Parallel Analysis}
\usage{
ggscree.plot(Df, title = NULL, use = "complete.obs", simu = 20,
  xlab = "Component", ylab = "Eigenvalue")
}
\arguments{
\item{Df}{a data frame containing the items}

\item{title}{title of the plot}

\item{use}{handling of missig values}

\item{simu}{number of simulations}

\item{xlab}{x-axis label of the plot}

\item{ylab}{y-axis label of the plot}
}
\description{
Performs parallel analysis of a set of items.
}
\details{
Adapted from \code{psy}-package (Bruno Falissard).
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

