% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemanalysis.R
\docType{package}
\name{iana-package}
\alias{iana-package}
\alias{iana}
\title{GUI for Item Analysis and Scale Construction}
\description{
Iana is a browser-based GUI for classical item and test analysis, factor analysis, and item response modeling with a focus on items with an ordered-category response format.
}
\details{
Iana is a browser-based graphical user interface to R functions for the psychometric analysis of questionnaires and tests with an ordinal response format. Iana tries to integrate the essential statistical analysis steps into a convenient interface. Iana covers classical item and test analysis (reliability, item discrimation), dimensionality tests (parallel analysis and MAP test), principal components and exploratory factor analysis (including factor analysis based on polychoric correlations), one-factor confirmatory analysis, and item response models (partial credit model). Graphical output includes histograms of item and test scores, empirical item characteric curves, and person-item maps, among others.

Iana is based on the \href{https://shiny.posit.co}{Shiny Web Framework for R}, which allows a fast bidirectional communication between the browser and R. Iana is "reactive", meaning that its output is instantly updated if any of the inputs (e.g., the selected variables) are changed by the user. This makes it easy to compare the impact of item selection and/or modeling options on the results. Iana keeps track of the R commands it constructs as a response to user input, so the analysis steps are documented and can be replicated. Iana comes with some built-in data sets, with which the interface can be tested. However, other data can easily be used.

The basic usage is documented in \code{\link{runiana}}.

\tabular{ll}{
Package: \tab iana\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2019-06-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\examples{
\dontrun{runiana()}
}
\references{
Shiny web framework for R: \url{https://shiny.posit.co}
}
\author{
Michael Hock (\email{michael.hock@uni-bamberg.de})
}
\keyword{package}
