% Generated by roxygen2 (4.0.2): do not edit by hand
\name{meansub}
\alias{meansub}
\title{Mean substitution of missing values}
\usage{
meansub(x, imp.value = NULL)
}
\arguments{
\item{x}{a data frame}

\item{imp.value}{imputed value. If NULL (default), the rounded mean is imputed.}
}
\value{
data frame frame with imputed values
}
\description{
Replaces missing values in numeric variables of a data frame by
rounded means or a specific value.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

