% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overview}
\alias{overview}
\title{Overview of a Data Frame}
\usage{
overview(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a data frame containing abbreviated name, class (first component), start column, end column, number of columns occupied for each variable, and variable range.
}
\description{
\code{overview} lists the columns in a data frame in a concise manner by stripping off trailing numbers from the column names. This is useful for getting an overview of questionnaire data where names of items belonging to a scale share the same stem and differ by an appended number.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

