% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{parallelAnalysis}
\alias{parallelAnalysis}
\title{Parallel analysis}
\usage{
parallelAnalysis(
  x,
  fm = "ml",
  cor = "cor",
  n.factors = NULL,
  sim = TRUE,
  onlyFA = FALSE
)
}
\arguments{
\item{x}{a data frame containing the variables (items) to be analyzed}

\item{fm}{string specifying the factoring method ("ml", "minres", ...)}

\item{cor}{string specifying how the correlations are found: "cor" (correlation), "cov" (covariance), "tet" (tetrachoric correlation), "poly" (polychoric correlation)}

\item{n.factors}{integer specifying the number of factors to plot; if NULL plot all factors}

\item{sim}{logical; if TRUE, sample from normal distribution, otherwise resample from the data}

\item{onlyFA}{logical; if TRUE, show only factor analysis results, if FALSE show also results from principal components}
}
\value{
a ggplot2 object
}
\description{
Performs a parallel analysis for the variables (usually items) in a data frame.
}
\author{
Computation: William Revelle, see \code{\link{psych}}, specifically \code{\link{fa.parallel}}.

Plot: Michael Hock, \email{michael.hock@uni-bamberg.de}
}
