% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{plot.mapTest}
\alias{plot.mapTest}
\title{Plot mapTest Object}
\usage{
\method{plot}{mapTest}(x, ...)
}
\arguments{
\item{x}{mapTest object}

\item{...}{ignored}
}
\description{
Plot mapTest object, see \code{\link{mapTest}}.
}

