% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{basicDescr}
\alias{basicDescr}
\title{Basic Descriptive Statistics}
\usage{
basicDescr(x)
}
\arguments{
\item{x}{a numeric variable}
}
\value{
a data frame
}
\description{
Return mean, standard deviation, skewness, and kurtosis of a variable. Missing values are removed before the computation.
}
