% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{frequencies}
\alias{frequencies}
\title{Frequencies}
\usage{
frequencies(x)
}
\arguments{
\item{x}{a data frame}
}
\description{
Returns a data frame representing a table with frequency counts of the
unique values of the variables in x. If present, item stems are appended
as the last row of the table.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

