% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{getDataFrames}
\alias{getDataFrames}
\title{Data Frames in Global Environment}
\usage{
getDataFrames()
}
\value{
A character vector with the names of the data frames
}
\description{
Returns of a vector with the names of the data frames present in the
global environment.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

