% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{getItemText}
\alias{getItemText}
\title{Return the Labels Associated with Items}
\usage{
getItemText(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a character vector of class 'itemText' containing the labels of the items
}
\description{
Return the labels associated with items in a data frame.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

