% Generated by roxygen2 (4.0.2): do not edit by hand
\name{overview}
\alias{overview}
\title{Overview of a Data Frame}
\usage{
overview(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a data frame containing abbreviated name (basename), class (first component), start column, end column, number of columns occupied by the variables with the same basename, and variable range.
}
\description{
\code{overview} lists the variables in a data frame in a concise manner by stripping off trailing numbers from the variable names (creating a basename) and combining variables with the same basename into one output row.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

