% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itemanalysis.R
\name{tryPrintExpr}
\alias{tryPrintExpr}
\title{Try to Print Results of a Command}
\usage{
tryPrintExpr(x)
}
\arguments{
\item{x}{an expression}
}
\description{
Try to print the results of a command, catching errors and warnings
with \code{\link{tryCatch.W.E}}. Warnings are appended to the
results.
}
\author{
Michael Hock \email{michael.hock@uni-bamberg.de}
}

